requirements.txt
netCDF4==1.5.5
numpy==1.20.0rc1
matplotlib==3.1.3
python-dateutil==2.8.1
pyhdf==0.10.2
coloredlogs==14.0

# In addition to these Python requirements, WIMSOFT is required to use these tool
# These tools are just an easy integration of the many WIMSOFT CLI